IOMAD: clear attendees needs to have a confirmation page

+++ mod/trainingevent/view.php
@@ -42,6 +42,7 @@ $current = optional_param('current', 0, PARAM_INTEGER);
 $chosen = optional_param('chosenevent', 0, PARAM_INTEGER);
 $action = optional_param('action', null, PARAM_ALPHA);
 $booking = optional_param('booking', null, PARAM_ALPHA);
+$confirm      = optional_param('confirm', '', PARAM_ALPHANUM);
 
 if (! $cm = get_coursemodule_from_id('trainingevent', $id)) {
     throw new moodle_exception('invalidcoursemodule');
@@ -227,7 +228,7 @@ if (!$event = $DB->get_record('trainingevent', array('id' => $cm->instance))) {
                                 foreach ($otherevents as $otherevent) {
                                     $DB->delete_records('trainingevent_users', ['trainingeventid' => $otherevent->id, 'userid' => $USER->id, 'waitlisted' => 1]);
                                 }
-                            } 
+                            }
                         }
                     } else {
                         $res = $DB->insert_record('trainingevent_users', array('trainingeventid' => $event->id, 'userid' => $USER->id));
@@ -1029,8 +1030,19 @@ if (!$event = $DB->get_record('trainingevent', array('id' => $cm->instance))) {
             }
         }
         if ($action == 'reset') {
-            if (has_capability('mod/trainingevent:resetattendees', $context)) {
-                $DB->delete_records('trainingevent_users', array('trainingeventid' => $event->id, 'waitlisted' => 0));
+            if ($confirm != md5($action)) {
+                echo $OUTPUT->header();
+                echo $OUTPUT->heading(get_string('resetattending', 'trainingevent'));
+                $optionsyes = ['id' => $id, 'action' => 'reset', 'confirm' => md5($action), 'sesskey' => sesskey()];
+                echo $OUTPUT->confirm(get_string('resetattendingfull', 'trainingevent'),
+                                                  new moodle_url('/mod/trainingevent/view.php', $optionsyes),
+                                                                 new moodle_url('/mod/trainingevent/view.php', ['id' => $id]));
+                echo $OUTPUT->footer();
+                die;
+            } else {
+                if (has_capability('mod/trainingevent:resetattendees', $context)) {
+                    $DB->delete_records('trainingevent_users', array('trainingeventid' => $event->id, 'waitlisted' => 0));
+                }
             }
         }
         if ($action == 'grade' && !empty($usergradeusers)) {
